# backend/openapi.yaml
openapi: 3.0.3
info:
  title: GADAVIRAL API
  description: |
    Ga & Dangme online social community API. One backend for Website, Android and Windows.
    Auth: JWT bearer access tokens (15 min) + rotating refresh tokens (30 days).
    Google sign-in uses genuine OpenID Connect ID tokens verified against Google JWKS.
    Authentication emails are sent from GADAVIRAL <admin@gadaviral.com>.
    Full endpoint reference: docs/API.md
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /posts:
    get: { tags: [Posts], summary: 'Feed (recent|following|recommended), paginated', security: [{ bearerAuth: [] }], responses: { '200': { description: OK } } }
    post: { tags: [Posts], summary: Create post (verified email; moderated), security: [{ bearerAuth: [] }], responses: { '201': { description: OK }, '403': { description: EMAIL_NOT_VERIFIED } } }
  /posts/media:
    post: { tags: [Posts], summary: 'Create media post (multipart, ≤6 files)', security: [{ bearerAuth: [] }], responses: { '201': { description: OK } } }
  '/posts/{id}':
    get: { tags: [Posts], summary: Post detail (records view), parameters: [{ '$ref': '#/components/parameters/Id' }], responses: { '200': { description: OK } } }
    patch: { tags: [Posts], summary: Edit own post, security: [{ bearerAuth: [] }], parameters: [{ '$ref': '#/components/parameters/Id' }], responses: { '200': { description: OK } } }
    delete: { tags: [Posts], summary: Delete/remove post, security: [{ bearerAuth: [] }], parameters: [{ '$ref': '#/components/parameters/Id' }], responses: { '200': { description: OK } } }
  '/posts/{id}/react':
    put: { tags: [Posts], summary: React (unique per user+post), security: [{ bearerAuth: [] }], parameters: [{ '$ref': '#/components/parameters/Id' }], responses: { '200': { description: OK } } }
    delete: { tags: [Posts], summary: Remove reaction, security: [{ bearerAuth: [] }], parameters: [{ '$ref': '#/components/parameters/Id' }], responses: { '200': { description: OK } } }
  '/posts/{id}/comments':
    get: { tags: [Posts], summary: Comments list, parameters: [{ '$ref': '#/components/parameters/Id' }], responses: { '200': { description: OK } } }
    post: { tags: [Posts], summary: 'Comment / reply (parentCommentId)', security: [{ bearerAuth: [] }], parameters: [{ '$ref': '#/components/parameters/Id' }], responses: { '201': { description: OK } } }
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }